Move MesonException from coredata to mesonlib.
