+++ mesonbuild/backend/xcodebackend.py
@@ -17,7 +17,7 @@ from .. import build
 from .. import mesonlib
 import uuid, os, sys
 
-from ..coredata import MesonException
+from ..mesonlib import MesonException
 
 class XCodeBackend(backends.Backend):
     def __init__(self, build):
